qlik-sales-ai-dashboard/qlik-sales-ai-dashboard/requirements.txt

**2. `requirements.txt`**
```txt
pandas
openai>=1.0.0
python-dotenv